Moved the student add run to a dialog popup. #1385
