+++ src/main/webapp/site/src/app/student/add-project-dialog/add-project-dialog.component.scss
@@ -1,0 +1,3 @@
+.add-run-input {
+  margin: 10px;
+}
