+++ src/main/webapp/site/src/app/student/student.service.ts
@@ -6,6 +6,8 @@ import { of } from 'rxjs/observable/of';
 
 import { RunInfo } from './run-info';
 import { StudentRun } from './student-run';
+import { Run } from "../domain/run";
+import { Subject } from "rxjs/Subject";
 
 @Injectable()
 export class StudentService {
@@ -13,6 +15,9 @@ export class StudentService {
   private runsUrl = 'api/student/runs';
   private runInfoUrl = 'api/student/run/info';
   private addRunUrl = 'api/student/run/register';
+
+  private newRunSource = new Subject<StudentRun>();
+  newRunSource$ = this.newRunSource.asObservable();
 
   constructor(private http: HttpClient) { }
 
@@ -60,5 +65,9 @@ export class StudentService {
 
   private log(message: string) {
     console.log('StudentService: ' + message);
+  }
+
+  addNewProject(run: StudentRun) {
+    this.newRunSource.next(run);
   }
 }
